requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.1
python-multipart==0.0.20
jinja2==3.1.5
pytesseract==0.3.13
pillow==11.0.0
pandas==2.2.3
openpyxl==3.1.5
langchain==0.3.13
langchain-huggingface==0.1.2
langchain-community==0.3.13
chromadb==0.5.23
sentence-transformers==3.3.1
